+++ compiler/src/main/scala/dxCompiler/Main.scala
@@ -135,6 +135,7 @@ object Main {
       "projectWideReuse" -> FlagOptionSpec.default,
       "reorg" -> FlagOptionSpec.default,
       "runtimeDebugLevel" -> IntOptionSpec.one.copy(choices = Vector(0, 1, 2)),
+      "separateOutputs" -> FlagOptionSpec.default,
       "streamFiles" -> StreamFilesOptionSpec,
       "streamAllFiles" -> FlagOptionSpec.default,
       "scatterChunkSize" -> IntOptionSpec.one,
@@ -440,6 +441,7 @@ object Main {
           leaveWorkflowsOpen,
           locked,
           projectWideReuse,
+          separateOutputs,
           streamAllFiles
       ) = Vector(
           "archive",
@@ -447,6 +449,7 @@ object Main {
           "leaveWorkflowsOpen",
           "locked",
           "projectWideReuse",
+          "separateOutputs",
           "streamAllFiles"
       ).map(options.getFlag(_))
       val streamFiles = options.getValue[StreamFiles.StreamFiles]("streamFiles") match {
@@ -466,6 +469,7 @@ object Main {
           leaveWorkflowsOpen,
           locked,
           projectWideReuse,
+          separateOutputs,
           streamFiles,
           useManifests,
           fileResolver
@@ -778,6 +782,8 @@ object Main {
         |                             How much debug information to write to the job log at runtime.
         |                             Log the minimum (0), intermediate (1, the default), or all 
         |                             debug information (2, for internal debugging).
+        |      -separateOutputs       Store the output files of each call in a separate folder. The
+        |                             default behavior is to put all outputs in the same folder.                             
         |      -streamFiles [all,none,perfile]
         |                             Whether to mount all files with dxfuse (do not use the 
         |                             download agent), to mount no files with dxfuse (only use