+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## in dev
+
+* Adds `-separateOutputs` option to store output from each call in a separate folder
+
+## 2.2.1 17-02-2021
+
+* Fix: native tasks now use instance type from native app/let, not wrapper task
+
 ## 2.2.0 12-02-2021
 
 * DxNI handles native app(let)s with optional non-file object inputs